chore: moved api-client-manager file
